Start working on updating user section
