+++ front-end/src/lib/models/section.ts
@@ -1,0 +1,4 @@
+export interface Section {
+  id: number;
+  name: string;
+}
